requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-core
huggingface-hub
sentence-transformers
faiss-cpu
PyPDF2
python-dotenv